+++ lib/hightop/utils.rb
@@ -8,7 +8,7 @@ module Hightop
         unless column.is_a?(Symbol) || column.is_a?(Arel::Nodes::SqlLiteral)
           column = column.to_s
           unless /\A\w+(\.\w+)?\z/i.match(column)
-            warn "[hightop] Non-attribute argument: #{column}. Use Arel.sql() for known-safe values. This will raise an error in Hightop 0.3.0"
+            raise ActiveRecord::UnknownAttributeReference, "Query method called with non-attribute argument(s): #{column.inspect}. Use Arel.sql() for known-safe values."
           end
         end
         column